Eliminated need for buildallmatches.sh… program now looks directly at ancestry finder files
The program extracts the 23andMe user's name from the filename using
the default file name provided by 23andMe; to wit
"ancestry_finder_{name}_YYYYMMDD.csv" where name is the individual's
name and YYYYMMDD is the date. We strip off everything but {name}. If
the file's name is structured differently, it will cause problems.