Add subs for normalizing text
The filenames on the files generated by 23andMe do not contain
diacritics, but within the files, the names may. This change just
handles more diacritics. It imposes a performance cost. I found online
a better way of doing this—by normalizing the strings to unicode and
stripping off the diacritics, but I couldn’t get it to work. Left it
here for future use.